requirements.txt
# FastAPI + ASGI Server (Web Scanner)
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Google Cloud APIs (mantido para OCR)
google-cloud-vision==3.4.5
google-auth==2.29.0
google-generativeai==0.3.2

# Environment & Security
python-dotenv==1.0.0

# Image Processing & OCR (compatível com Python 3.13)
Pillow==10.4.0
numpy==1.26.4  # Fix ABI do OpenCV (evita NumPy 2.x)
opencv-python-headless==4.8.1.78  # Para processamento de imagens sem libGL
pyzbar==0.1.9  # Decodificação de códigos de barras com ZBar

# Telegram Bot (Multi-Entregador)
python-telegram-bot==22.5
# httpx será instalado automaticamente como dependência do python-telegram-bot

# WebSocket Dashboard (Real-time)
aiohttp==3.9.1
aiohttp-cors==0.7.0

# Excel Parser (Shopee romaneios)
openpyxl==3.1.2

# Parsers de Romaneios
pdfplumber==0.11.0  # PDF com texto extraível
PyPDF2==3.0.1  # Alternativa para PDF
pytesseract==0.3.10  # OCR para PDFs escaneados (requer Tesseract instalado)
pdf2image==1.17.0  # Converte PDF em imagem para OCR

# Exportação e Dashboard
reportlab==4.0.7  # Geração de PDF
requests==2.31.0  # HTTP client para API Banco Inter

# Banco de Dados (PostgreSQL para persistência no Railway)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter
alembic==1.13.1  # Database migrations
python-multipart>=0.0.9
